Rename files to reinforce operator terminology.
