+++ Tests/GauntletTests/Operators/CollectionOperatorsTests.swift
@@ -1,5 +1,5 @@
 //
-//  CollectionAssertionsTests.swift
+//  CollectionOperatorsTests.swift
 //
 //  MIT License
 //
